+++ src/D2L.CodeStyle.Analyzers/D2L.CodeStyle.Analyzers.csproj
@@ -36,8 +36,8 @@
     <Compile Include="ApiUsage\DangerousMemberUsages\DangerousMemberUsagesAnalyzer.cs" />
     <Compile Include="ApiUsage\DangerousMemberUsages\DangerousMethods.cs" />
     <Compile Include="ApiUsage\DangerousMemberUsages\DangerousProperties.cs" />
-    <Compile Include="ApiUsage\GenericTypeArguments\GenericTypeArgumentsBlacklistAnalyzer.cs" />
-    <Compile Include="ApiUsage\GenericTypeArguments\GenericTypeArgumentsBlacklist.cs" />
+    <Compile Include="ApiUsage\Events\EventHandlersBlacklistAnalyzer.cs" />
+    <Compile Include="ApiUsage\Events\EventHandlersBlacklist.cs" />
     <Compile Include="Immutability\AddImmutableAttributeCodeFix.cs" />
     <Compile Include="Immutability\ExplicitImmutableAttributeTransitivityAnalyzer.cs" />
     <Compile Include="Immutability\ImmutableGenericArgument.cs" />
