+++ tests/D2L.CodeStyle.Analyzers.Test/D2L.CodeStyle.Analyzers.Tests.csproj
@@ -54,7 +54,7 @@
     <EmbeddedResource Include="Specs\ImmutableGenericDeclarationAnalyzer.cs" />
     <EmbeddedResource Include="Specs\DangerousMemberUsagesAnalyzer.DangerousProperties.cs" />
     <EmbeddedResource Include="Specs\ExplicitImmutableAttributeTransitivityAnalyzer.cs" />
-    <EmbeddedResource Include="Specs\GenericTypeArgumentsBlacklistAnalyzer.cs" />
+    <EmbeddedResource Include="Specs\EventHandlersBlacklistAnalyzer.cs" />
     <Compile Include="TestBase.cs" />
     <Compile Include="Extensions\TypeSymbolExtensionsTests.cs" />
     <Compile Include="ApiUsage\NotNullAnalyzerTests.cs" />
